+++ CHANGELOG.md
@@ -4,9 +4,19 @@ ADDED:
 
 CHANGED:
 
+- `Link::createFromArray()` performs type checks before instantiating a `Link` object
+- The "jsonapi" member is always present when using `Document::toArray()` 
+
 REMOVED:
 
+- `Links::hasLinks()` method
+- `JsonApi::hasJsonApi()` method
+
 FIXED:
+
+- The "jsonapi" member was incorrectly recognized as `jsonApi`
+- If the "jsonapi" member doesn't define the `version`, "1.0" will be the default now as per the spec
+- Invocation of `Links::hasLink()` and thus `Links::hasSelf()` etc. methods resulted in an infinite loop
 
 ## 0.7.0 - 2016-12-19
 
@@ -21,7 +31,7 @@ CHANGED:
 
 - Added `JsonApiResource::setToOneRelationship()` instead of `JsonApiResource::setToOneResourceIdentifier()`
 - Added `JsonApiResource::setToManyRelationship()` instead of `JsonApiResource::addToManyResourceIdentifier()`
-- A `Links` member won't be present in the `data` member when it is empty when using `Document::toArray()`
+- A "links" member won't be present in the "data" member when it is empty when using `Document::toArray()`
 - Renamed several methods of `ResourceObjects` (removed `get` prefixes)
 - Renamed `JsonApiRelationshipInterface` class to `RelationshipInterface`
 - Renamed `JsonApiResource` class to `ResourceObject`
@@ -35,10 +45,10 @@ FIXED:
 - `JsonApiRequestBuilder::setPort()` didn't do anything
 - `JsonApiRequestBuilder::setJsonApi*()` methods didn't work as intended
 - `JsonApiRequestBuilder::update()` will now set the request method to "PATCH" instead of "UPDATE"
-- Error objects are correctly listed under the `errors` member when using `Document::toArray()`
-- Correctly transforming `included` member when using `Document::toArray()`
-- `ResourceObjects::isSingleResourceDocument()` `Document::isResourceDocument()` returned wrong value When `data` member was null 
-- `ResourceObjects::hasAnyPrimaryResources()` returned wrong value When `data` member was null 
+- Error objects are correctly listed under the "errors" member when using `Document::toArray()`
+- Correctly transforming "included" member when using `Document::toArray()`
+- `ResourceObjects::isSingleResourceDocument()` `Document::isResourceDocument()` returned wrong value when "data" member was null 
+- `ResourceObjects::hasAnyPrimaryResources()` returned wrong value When "data" member was null 
 
 ## 0.6.0 - 2016-12-18